server/requirements.txt
flask
flask-cors
psycopg2-binary
python-dotenv
bcrypt
pyjwt
pillow
pytesseract